Add Web3 Ticketing (#378)
Version is 0.0.2

(This will be invalid because the icon is missing, will follow up in
another commit)

Closes #377 

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->