Add screenshots field (#505)
This adds support for a screenshots field in the registry. Snaps may
have three screenshots, with a size of 960x540. Screenshots should be
placed in `src/images/<snapId>`, i.e.,
`src/images/@organisation/snap-name`, and should be named `1.png`,
`2.png`, `3.png` (or `.jpe?g`). They can then be added to the registry
by adding `screenshots` to the metadata, with the path to each file as
array items.

```json
{
  "verifiedSnaps": {
    "npm:@organisation/snap-name": {
      "metadata": {
        "screenshots": [
          "./images/@organisation/snap-name/1.png",
          "./images/@organisation/snap-name/2.png",
          "./images/@organisation/snap-name/3.png"
        ]
      }
    }
  }
}
```