Utility to query Snaps for conditions. (#520)
This PR is for a Snap Query Utility.
Occasionally we have the need to query for Snaps that meet a condition
or criteria - the main use-case being notifying a Snap Builder of an
upcoming change or other reason.

The PR has 2 code files:

- `get-manifests.ts` : The code in this file is used to download the
manifest files of all Snaps. This will get called the first time or when
a refresh of the manifests are needed. All the Snap manifests are then
saved to a local JSON file (`./data.json`). Local storage is to avoid
fetching Snap manifests every time we need to run a query.
- `query.ts` : This file contains the query logic that we are interested
in, and works on the `./data.json` file.

Usage:
```
yarn run query-snaps

options:
--force-download : This forces a download of all Snap manifests (refresh).
```

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>